+++ src/Frontend/Dfe.Complete/Pages/Projects/Team/ProjectsUnassigned/AllProjectsUnassignedForTeamModel.cshtml.cs
@@ -9,6 +9,7 @@ using Dfe.Complete.Pages.Pagination;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Dfe.Complete.Domain.Constants;
+using Microsoft.AspNetCore.Mvc;
 
 namespace Dfe.Complete.Pages.Projects.Team.ProjectsUnassigned;
 
@@ -18,7 +19,7 @@ public class AllProjectsUnassignedForTeamModel(ISender sender) : YourTeamProject
     public List<ListAllProjectsResultModel> Projects { get; set; } = [];
     public bool UserTeamIsRegionalDeliveryOfficer { get; set; }
 
-    public async Task OnGet()
+    public async Task<IActionResult> OnGet()
     {
         ViewData[TabNavigationModel.ViewDataKey] = YourTeamProjectsTabNavigationModel;
 
@@ -56,6 +57,9 @@ public class AllProjectsUnassignedForTeamModel(ISender sender) : YourTeamProject
         }
 
         Pagination = new PaginationModel(RouteConstants.TeamProjectsUnassigned, PageNumber, recordCount, PageSize);
+
+        var hasPageFound = HasPageFound(Pagination.IsOutOfRangePage);
+        return hasPageFound ?? Page();
     }
 
     public async Task OnGetMovePage()