+++ src/Tests/Dfe.Complete.Tests/Pages/Pagination/PaginationModelTests.cs
@@ -178,4 +178,38 @@ public class PaginationModelTests
         // Assert
         Assert.Equal("custom-previous-page", model.PreviousButtonId);
     }
+
+    [Theory]
+    [InlineData(1, false)]
+    [InlineData(200, true)]
+    public void IsOutOfRangePage_ShouldReturnCorrectResult(int pageNumber, bool result)
+    {
+        // Arrange
+        var url = "https://example.com"; 
+        var recordCount = 10;
+        var pageSize = 5;
+        var elementIdPrefix = "custom-";
+
+        // Act
+        var model = new PaginationModel(url, pageNumber, recordCount, pageSize, elementIdPrefix);
+
+        // Assert
+        Assert.Equal(result, model.IsOutOfRangePage);
+    }
+
+    [Theory]
+    [InlineData("https://example.com")]
+    [InlineData("https://example.com?q=test")]
+    public void SetUrl_FormatesCorrectURL(string url)
+    {
+        // Arrange
+        var pageNumber = 1;
+        var expectedUrl = $"{url}{(url.Contains('?') ? "&" : "?")}page={pageNumber}";
+
+        // Act 
+        var result = PaginationModel.SetUrl(url, pageNumber);
+
+        // Assert
+        Assert.Equal(expectedUrl, result);
+    }
 }