+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Security in case of vulnerabilities.
 ### Fixed  
 - Updated pagination query parameter from `pageNumber` to `page` to match Ruby app
 - Optimised queries behind the "By local authority" page
+- Show a `Page Not Found` error if the requested page number exceeds the total number of available pages.
 - Fixed unable to set cookies issue if request is coming from ruby app
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-05-30.320...main) for everything awaiting release